menu and sound files
Added a bunch of sound files to assets, which is a bunch of mp3's (we
can add or remove to this archive if needed). Also, we now have a menu
class in 'menu.lua' that can supply a start screen, a pause screen. and
a game over screen. The API should be safe and easy to use, so we can
easily maintain the code and add more functionality without affecting
any of our current code in 'main.lua'